requirements.txt
pyserial
pyFirmata
unittest2
mock


